Rolling in Chat Encryption
I refactored the handshake process so that keys are generated and public keys are provided to the server. The server
saves these keys in memory for use later when they will be provided to recipients as part of the client to client
handshake that's not implemented yet. See the README for details of which features are implemented and which features
still require more work.

As far as my simple testing, chat communication should still work as it did before but now using the new registration
logic that involves using data from the client to generate the UUID, creating and storing the public key for a client.

Error handling is still a big TODO.